chore: restructure module and remove uneeded parser funcs
