+++ cmd/infracost/testdata/breakdown_terraform_directory_with_modules/breakdown_terraform_directory_with_modules.golden
@@ -3,7 +3,7 @@ Project: infracost/infracost/examples/terraformwithmodules
  Name                                                           Monthly Qty  Unit                        Monthly Cost 
                                                                                                                       
  aws_instance.web_app                                                                                                 
- ├─ Instance usage (Linux/UNIX, on-demand, m5.4xlarge)                  730  hours                            $560.64 
+ ├─ Instance usage (Linux/UNIX, on-demand, m5.2xlarge)                  730  hours                            $280.32 
  ├─ root_block_device                                                                                                 
  │  └─ Storage (general purpose SSD, gp2)                                50  GB                                 $5.00 
  └─ ebs_block_device[0]                                                                                               
@@ -43,7 +43,7 @@ Project: infracost/infracost/examples/terraformwithmodules
  └─ ebs_block_device[0]                                                                                               
     └─ Storage (general purpose SSD, gp2)                             1,000  GB                               $100.00 
                                                                                                                       
- OVERALL TOTAL                                                                                              $4,603.48 
+ OVERALL TOTAL                                                                                              $4,323.16 
 ──────────────────────────────────
 6 cloud resources were detected, rerun with --show-skipped to see details:
 ∙ 6 were estimated, 6 include usage-based costs, see https://infracost.io/usage-file
