+++ examples/terraformwithmodules/main.tf
@@ -30,9 +30,17 @@ module "big_app_gp2" {
   volume_type   = "gp2"
 }
 
+resource "aws_lambda_function" "hello_world" {
+  function_name = "hello_world"
+  role          = "arn:aws:lambda:us-east-1:account-id:resource-id"
+  handler       = "exports.test"
+  runtime       = "nodejs12.x"
+  memory_size   = var.memory_size
+}
+
 resource "aws_instance" "web_app" {
   ami           = "ami-674cbc1e"
-  instance_type = "m5.4xlarge" # <<<<< Try changing this to m5.8xlarge to compare the costs
+  instance_type = var.instance_type
 
   root_block_device {
     volume_size = 50
@@ -40,16 +48,20 @@ resource "aws_instance" "web_app" {
 
   ebs_block_device {
     device_name = "my_data"
-    volume_type = "io1" # <<<<< Try changing this to gp2 to compare costs
+    volume_type = var.volume_type  # <<<<< Try changing this to gp2 to compare costs
     volume_size = 1000
     iops        = 800
   }
 }
 
-resource "aws_lambda_function" "hello_world" {
-  function_name = "hello_world"
-  role          = "arn:aws:lambda:us-east-1:account-id:resource-id"
-  handler       = "exports.test"
-  runtime       = "nodejs12.x"
-  memory_size   = 1024 # <<<<< Try changing this to 512 to compare costs
+variable "memory_size" {
+  default = 1024
+}
+
+variable "instance_type" {
+  default = "m5.2xlarge"
+}
+
+variable "volume_type" {
+  default = "io1"
 }
