+++ examples/terraformwithmodules/modules/instance/main.tf
@@ -5,6 +5,7 @@ variable "instance_type" {
 variable "volume_type" {
   default = ""
 }
+
 resource "aws_instance" "web_app" {
   ami           = "ami-674cbc1e"
   instance_type = var.instance_type
