+++ internal/hcl/attribute.go
@@ -1,4 +1,4 @@
-package block
+package hcl
 
 import (
 	"fmt"
