--- internal/hcl/block/schema.go
@@ -1,38 +1,0 @@
-package block
-
-import "github.com/hashicorp/hcl/v2"
-
-var TerraformSchema012 = &hcl.BodySchema{
-	Blocks: []hcl.BlockHeaderSchema{
-		{
-			Type: "terraform",
-		},
-		{
-			Type:       "provider",
-			LabelNames: []string{"name"},
-		},
-		{
-			Type:       "variable",
-			LabelNames: []string{"name"},
-		},
-		{
-			Type: "locals",
-		},
-		{
-			Type:       "output",
-			LabelNames: []string{"name"},
-		},
-		{
-			Type:       "module",
-			LabelNames: []string{"name"},
-		},
-		{
-			Type:       "resource",
-			LabelNames: []string{"type", "name"},
-		},
-		{
-			Type:       "data",
-			LabelNames: []string{"type", "name"},
-		},
-	},
-}
