+++ internal/hcl/context.go
@@ -1,4 +1,4 @@
-package block
+package hcl
 
 import (
 	"strings"
@@ -80,12 +80,16 @@ func (c *Context) Set(val cty.Value, parts ...string) {
 	c.ctx.Variables[parts[0]] = v
 }
 
-func mergeVars(src cty.Value, parts []string, value cty.Value) cty.Value {
+func isValidCtyObject(src cty.Value) bool {
+	return src.IsKnown() && src.Type().IsObjectType() && !src.IsNull() && src.LengthInt() > 0
+}
 
+func mergeVars(src cty.Value, parts []string, value cty.Value) cty.Value {
 	if len(parts) == 0 {
-		if value.IsKnown() && value.Type().IsObjectType() && !value.IsNull() && value.LengthInt() > 0 && src.IsKnown() && src.Type().IsObjectType() && !src.IsNull() && src.LengthInt() > 0 {
+		if isValidCtyObject(value) && isValidCtyObject(src) {
 			return mergeObjects(src, value)
 		}
+
 		return value
 	}
 
@@ -111,10 +115,11 @@ func mergeObjects(a cty.Value, b cty.Value) cty.Value {
 	for key, val := range a.AsValueMap() {
 		output[key] = val
 	}
+
 	for key, val := range b.AsValueMap() {
 		old, exists := output[key]
 
-		if exists && val.IsKnown() && val.Type().IsObjectType() && !val.IsNull() && val.LengthInt() > 0 && old.IsKnown() && old.Type().IsObjectType() && !old.IsNull() && old.LengthInt() > 0 {
+		if exists && isValidCtyObject(val) && isValidCtyObject(old) {
 			output[key] = mergeObjects(val, old)
 			continue
 		}
