+++ internal/hcl/functions.go
@@ -1,4 +1,4 @@
-package parser
+package hcl
 
 import (
 	"github.com/hashicorp/hcl/v2/ext/tryfunc"
