--- internal/hcl/parser/load_dir.go
@@ -1,62 +1,0 @@
-package parser
-
-import (
-	"fmt"
-	"io/ioutil"
-	"os"
-	"path/filepath"
-	"strings"
-
-	"github.com/hashicorp/hcl/v2"
-	"github.com/hashicorp/hcl/v2/hclparse"
-)
-
-var knownFiles = make(map[string]struct{})
-
-func LoadDirectory(fullPath string, stopOnHCLError bool) ([]*hcl.File, error) {
-	hclParser := hclparse.NewParser()
-
-	fileInfos, err := ioutil.ReadDir(fullPath)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, info := range fileInfos {
-		if info.IsDir() {
-			continue
-		}
-
-		var parseFunc func(filename string) (*hcl.File, hcl.Diagnostics)
-		if strings.HasSuffix(info.Name(), ".tf") {
-			parseFunc = hclParser.ParseHCLFile
-		}
-
-		if strings.HasSuffix(info.Name(), ".tf.json") {
-			parseFunc = hclParser.ParseJSONFile
-		}
-
-		// this is not a file we can parse:
-		if parseFunc == nil {
-			continue
-		}
-
-		path := filepath.Join(fullPath, info.Name())
-		_, diag := parseFunc(path)
-		if diag != nil && diag.HasErrors() {
-			if stopOnHCLError {
-				return nil, diag
-			}
-			_, _ = fmt.Fprintf(os.Stderr, "WARNING: HCL error: %s\n", diag)
-			continue
-		}
-
-		knownFiles[path] = struct{}{}
-	}
-
-	var files []*hcl.File
-	for _, file := range hclParser.Files() {
-		files = append(files, file)
-	}
-
-	return files, nil
-}
