--- internal/hcl/parser/load_vars.go
@@ -1,51 +1,0 @@
-package parser
-
-import (
-	"fmt"
-	"io/ioutil"
-
-	"github.com/hashicorp/hcl/v2"
-	"github.com/hashicorp/hcl/v2/hclsyntax"
-	log "github.com/sirupsen/logrus"
-	"github.com/zclconf/go-cty/cty"
-)
-
-func LoadTFVars(filenames []string) (map[string]cty.Value, error) {
-	combinedVars := make(map[string]cty.Value)
-
-	for _, filename := range filenames {
-		vars, err := loadTFVars(filename)
-		if err != nil {
-			return nil, fmt.Errorf("failed to load the tfvars. %s", err.Error())
-		}
-		for k, v := range vars {
-			combinedVars[k] = v
-		}
-	}
-
-	return combinedVars, nil
-}
-
-func loadTFVars(filename string) (map[string]cty.Value, error) {
-	inputVars := make(map[string]cty.Value)
-
-	if filename == "" {
-		return inputVars, nil
-	}
-
-	log.Debugf("loading tfvars-file [%s]", filename)
-	src, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return nil, err
-	}
-
-	variableFile, _ := hclsyntax.ParseConfig(src, filename, hcl.Pos{Line: 1, Column: 1})
-	attrs, _ := variableFile.Body.JustAttributes()
-
-	for _, attr := range attrs {
-		log.Debugf("Setting '%s' from tfvars file at %s", attr.Name, filename)
-		inputVars[attr.Name], _ = attr.Expr.Value(&hcl.EvalContext{})
-	}
-
-	return inputVars, nil
-}
