+++ internal/hcl/reference_test.go
@@ -1,4 +1,4 @@
-package block
+package hcl
 
 import (
 	"testing"
