+++ internal/hcl/value_functions.go
@@ -1,4 +1,4 @@
-package block
+package hcl
 
 import (
 	"fmt"
